requirements.txt
amqp
arabic-reshaper
asgiref
asn1crypto
billiard
black
Brotli
celery
certifi
cffi
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
crispy-bootstrap5
cryptography
cssselect2
Deprecated
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-environ
django-extensions
django-filter
django-jet-reboot
django-model-utils
django-redis
djangorestframework
djangorestframework-simplejwt
EasyProcess
entrypoint2
gopay
hiredis
html5lib
idna
kombu
lxml
Markdown
MouseInfo
mss
mutagen
mypy-extensions
numpy
opencv-contrib-python
oscrypto
packaging
pathspec
Pillow
platformdirs
prompt_toolkit
psycopg2
psycopg2-binary
PyAutoGUI
pycparser
pycryptodomex
pydantic
PyGetWindow
pyHanko
pyhanko-certvalidator
PyJWT
PyMsgBox
pypdf
pyperclip
pypng
PyRect
pyscreenshot
PyScreeze
python-bidi
python-dateutil
python-decouple
pytweening
pytz
PyYAML
qrcode
redis
reportlab
requests
six
sqlparse
stripe
svglib
tinycss2
typing_extensions
tzdata
tzlocal
uritools
urllib3
vine
wcwidth
webencodings
websockets
whitenoise
wrapt
xhtml2pdf
yt-dlp
